Consolidate Fake data, and make it consistent across multiple calls.
